requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0
flasgger==0.9.7.1
instagrapi==2.0.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==21.2.0
pytest==8.1.1